backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-dotenv
httpx
groq
pandas
python-multipart
playwright
playwright-stealth
chromadb
PyPDF2
sentence-transformers
duckduckgo-search
facebook-business
google-ads
